+++ Extend/Controller/Admin/PaymentMain.php
@@ -13,12 +13,14 @@ use OxidEsales\Eshop\Core\Registry;
 
 use Wirecard\Oxid\Core\Helper;
 use Wirecard\Oxid\Core\PaymentMethodHelper;
+
 use Wirecard\Oxid\Extend\Model\Payment;
-use Wirecard\Oxid\Model\PayolutionBtwobPaymentMethod;
-use Wirecard\Oxid\Model\CreditCardPaymentMethod;
-use Wirecard\Oxid\Model\SofortPaymentMethod;
-use Wirecard\Oxid\Model\SepaDirectDebitPaymentMethod;
-use Wirecard\Oxid\Model\PayolutionInvoicePaymentMethod;
+
+use Wirecard\Oxid\Model\PaymentMethod\PayolutionBtwobPaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\CreditCardPaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\SofortPaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\SepaDirectDebitPaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\PayolutionInvoicePaymentMethod;
 
 use Wirecard\PaymentSdk\Config\Config;
 use Wirecard\PaymentSdk\TransactionService;
@@ -200,7 +202,7 @@ class PaymentMain extends PaymentMain_parent
         // the country code must be of the format two characters, underscore, two characters: "en_gb"
         // this format is checked by the regular expression below
         $sCountryCode = $aParams['oxpayments__wdoxidee_countrycode'];
-        return $aParams['oxpayments__oxid'] !== SofortPaymentMethod::getName(true)
+        return $aParams['oxpayments__oxid'] !== SofortPaymentMethod::getName()
             || (preg_match('/^[a-z]{2}_[a-z]{2}$/', $sCountryCode) === 1);
     }
 
@@ -216,7 +218,7 @@ class PaymentMain extends PaymentMain_parent
     private function _isCreditorIdValid($aParams)
     {
         $sCreditorId = $aParams['oxpayments__wdoxidee_creditorid'];
-        return $aParams['oxpayments__oxid'] !== SepaDirectDebitPaymentMethod::getName(true)
+        return $aParams['oxpayments__oxid'] !== SepaDirectDebitPaymentMethod::getName()
             || $this->_creditorIdValidation($sCreditorId);
     }
