+++ Model/PaymentMethod/GiropayPaymentMethod.php
@@ -7,28 +7,31 @@
  * https://github.com/wirecard/oxid-ee/blob/master/LICENSE
  */
 
-namespace Wirecard\Oxid\Model;
+namespace Wirecard\Oxid\Model\PaymentMethod;
+
+use OxidEsales\Eshop\Core\Registry;
 
 use Wirecard\PaymentSdk\Config\Config;
 use Wirecard\PaymentSdk\Config\PaymentMethodConfig;
-use Wirecard\PaymentSdk\Transaction\SofortTransaction;
+use Wirecard\PaymentSdk\Entity\BankAccount;
+use Wirecard\PaymentSdk\Transaction\GiropayTransaction;
 use Wirecard\PaymentSdk\Transaction\Transaction;
 
 use Wirecard\Oxid\Core\Helper;
 
 /**
- * Payment method implementation for Sofort.
+ * Payment method implementation for giropay.
  *
- * @since 1.0.0
+ * @since 1.2.0
  */
-class SofortPaymentMethod extends SepaCreditTransferPaymentMethod
+class GiropayPaymentMethod extends SepaCreditTransferPaymentMethod
 {
     /**
      * @inheritdoc
      *
-     * @since 1.0.0
+     * @since 1.2.0
      */
-    protected static $_sName = "sofortbanking";
+    protected static $_sName = 'giropay';
 
     /**
      * @inheritdoc
@@ -44,19 +47,19 @@ class SofortPaymentMethod extends SepaCreditTransferPaymentMethod
      *
      * @return Config
      *
-     * @since 1.0.0
+     * @since 1.2.0
      */
     public function getConfig()
     {
         $oConfig = parent::getConfig();
 
         $oPaymentMethodConfig = new PaymentMethodConfig(
-            SofortTransaction::NAME,
+            GiropayTransaction::NAME,
             $this->_oPayment->oxpayments__wdoxidee_maid->value,
             $this->_oPayment->oxpayments__wdoxidee_secret->value
         );
-        $oConfig->add($oPaymentMethodConfig);
 
+        $oConfig->add($oPaymentMethodConfig);
         return $oConfig;
     }
 
@@ -65,31 +68,29 @@ class SofortPaymentMethod extends SepaCreditTransferPaymentMethod
      *
      * @return Transaction
      *
-     * @since 1.0.0
+     * @since 1.2.0
      */
     public function getTransaction()
     {
-        return new SofortTransaction();
+        return new GiropayTransaction();
     }
 
     /**
-     * Sofort has a variable logo depending on the shop language
+     * @inheritdoc
      *
-     * @return string
+     * @param GiropayTransaction $oTransaction
+     * @param Order              $oOrder
      *
-     * @since 1.0.0
+     * @since 1.2.0
      */
-    public function getLogoPath()
+    public function addMandatoryTransactionData(&$oTransaction, $oOrder)
     {
-        $sLogoPath = $this->_oPayment->oxpayments__wdoxidee_logo->value;
-        $sCountryCode = $this->_oPayment->oxpayments__wdoxidee_countrycode->value;
-        $sLogoVariant = $this->_oPayment->oxpayments__wdoxidee_logovariant->value;
+        $oBankAccount = new BankAccount();
+        $oSession = Registry::getConfig()->getSession();
+        $aDynvalues = $oSession->getVariable('dynvalue');
 
-        return sprintf(
-            $sLogoPath,
-            $sCountryCode,
-            $sLogoVariant
-        );
+        $oBankAccount->setBic($aDynvalues['bic'] ?? '');
+        $oTransaction->setBankAccount($oBankAccount);
     }
 
     /**
@@ -97,19 +98,29 @@ class SofortPaymentMethod extends SepaCreditTransferPaymentMethod
      *
      * @return array
      *
-     * @since 1.0.0
+     * @since 1.2.0
      */
     public function getConfigFields()
     {
         $aAdditionalFields = [
-            'additionalInfo' => [
-                'type'        => 'select',
-                'field'       => 'oxpayments__wdoxidee_additional_info',
-                'options'     => [
-                    '1'       => Helper::translate('wd_yes'),
-                    '0'       => Helper::translate('wd_no'),
+            'descriptor' => [
+                'type'  => 'select',
+                'field' => 'oxpayments__wdoxidee_descriptor',
+                'options' => [
+                    '1' => Helper::translate('wd_yes'),
+                    '0' => Helper::translate('wd_no'),
                 ],
-                'title'       => Helper::translate('wd_config_additional_info'),
+                'title' => Helper::translate('wd_config_descriptor'),
+                'description' => Helper::translate('wd_config_descriptor_desc'),
+            ],
+            'additionalInfo' => [
+                'type'  => 'select',
+                'field' => 'oxpayments__wdoxidee_additional_info',
+                'options' => [
+                    '1' => Helper::translate('wd_yes'),
+                    '0' => Helper::translate('wd_no'),
+                ],
+                'title' => Helper::translate('wd_config_additional_info'),
                 'description' => Helper::translate('wd_config_additional_info_desc'),
             ],
             'deleteCanceledOrder' => [
@@ -132,26 +143,9 @@ class SofortPaymentMethod extends SepaCreditTransferPaymentMethod
                 'title' => Helper::translate('wd_config_delete_failure_order'),
                 'description' => Helper::translate('wd_config_delete_failure_order_desc'),
             ],
-            'countryCode' => [
-                'type'        => 'text',
-                'field'       => 'oxpayments__wdoxidee_countrycode',
-                'title'       => Helper::translate('wd_config_country_code'),
-                'description' => Helper::translate('wd_config_country_code_desc'),
-                'onchange'    => 'wdCheckCountryCode()',
-            ],
-            'logoType' => [
-                'type'        => 'select',
-                'field'       => 'oxpayments__wdoxidee_logovariant',
-                'options'     => [
-                    'standard'      => Helper::translate('wd_text_logo_variant_standard'),
-                    'descriptive'   => Helper::translate('wd_text_logo_variant_descriptive'),
-                ],
-                'title'       => Helper::translate('wd_config_logo_variant'),
-                'description' => Helper::translate('wd_config_logo_variant_desc'),
-            ],
         ];
 
-        return parent::getConfigFields() + $aAdditionalFields;
+        return array_merge(parent::getConfigFields(), $aAdditionalFields);
     }
 
     /**
@@ -159,13 +153,33 @@ class SofortPaymentMethod extends SepaCreditTransferPaymentMethod
      *
      * @return array
      *
-     * @since 1.0.0
+     * @since 1.2.0
+     */
+    public function getCheckoutFields()
+    {
+        return [
+            'bic' => [
+                'type' => 'text',
+                'title' => Helper::translate('wd_bic'),
+                'required' => true,
+            ],
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     *
+     * @return array
+     *
+     * @since 1.2.0
      */
     public function getPublicFieldNames()
     {
-        return array_merge(
-            parent::getPublicFieldNames(),
-            ['additionalInfo', 'countryCode', 'logoType', 'deleteCanceledOrder', 'deleteFailedOrder']
-        );
+        return array_merge(parent::getPublicFieldNames(), [
+            'descriptor',
+            'additionalInfo',
+            'deleteCanceledOrder',
+            'deleteFailedOrder',
+        ]);
     }
 }
