+++ Tests/Acceptance/SofortCheckoutTest.php
@@ -9,7 +9,7 @@
 
 namespace Wirecard\Oxid\Tests\Acceptance;
 
-use Wirecard\Oxid\Model\SofortPaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\SofortPaymentMethod;
 
 /**
  * Acceptance tests for the Sofort. checkout flow.
