+++ Tests/Unit/Model/IdealPaymentMethodTest.php
@@ -7,14 +7,12 @@
  * https://github.com/wirecard/oxid-ee/blob/master/LICENSE
  */
 
+use OxidEsales\Eshop\Core\Registry;
 use Wirecard\Oxid\Core\PaymentMethodHelper;
-use Wirecard\Oxid\Model\IdealPaymentMethod;
-
+use Wirecard\Oxid\Model\PaymentMethod\IdealPaymentMethod;
 use Wirecard\PaymentSdk\Config\Config;
 use Wirecard\PaymentSdk\Config\PaymentMethodConfig;
 use Wirecard\PaymentSdk\Transaction\IdealTransaction;
-
-use OxidEsales\Eshop\Core\Registry;
 
 class IdealPaymentMethodTest extends OxidEsales\TestingLibrary\UnitTestCase
 {
@@ -31,12 +29,10 @@ class IdealPaymentMethodTest extends OxidEsales\TestingLibrary\UnitTestCase
 
     public function testGetConfig()
     {
-        $oPayment = PaymentMethodHelper::getPaymentById(IdealPaymentMethod::getName(true));
-
         $oConfig = $this->_oPaymentMethod->getConfig();
 
         $this->assertInstanceOf(Config::class, $oConfig);
-        $this->assertInstanceOf(PaymentMethodConfig::class, $oConfig->get(IdealPaymentMethod::getName()));
+        $this->assertInstanceOf(PaymentMethodConfig::class, $oConfig->get('ideal'));
     }
 
     public function testGetTransaction()
@@ -48,21 +44,10 @@ class IdealPaymentMethodTest extends OxidEsales\TestingLibrary\UnitTestCase
         $this->assertInstanceOf(IdealTransaction::class, $this->_oPaymentMethod->getTransaction());
     }
 
-    /**
-     * @dataProvider getNameProvider
-     */
-    public function testGetName($bForOxid, $sExpected)
+    public function testGetName()
     {
-        $sName = IdealPaymentMethod::getName($bForOxid);
-        $this->assertEquals($sExpected, $sName);
-    }
-
-    public function getNameProvider()
-    {
-        return [
-            'for oxid' => [true, 'wdideal'],
-            'not for oxid' => [false, 'ideal'],
-        ];
+        $sName = IdealPaymentMethod::getName();
+        $this->assertEquals('wdideal', $sName);
     }
 
     public function testGetConfigFields()
