+++ Tests/Unit/Model/RatepayInvoicePaymentMethodTest.php
@@ -7,7 +7,7 @@
  * https://github.com/wirecard/oxid-ee/blob/master/LICENSE
  */
 
-use Wirecard\Oxid\Model\RatepayInvoicePaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\RatepayInvoicePaymentMethod;
 use Wirecard\Oxid\Extend\Model\Payment;
 use Wirecard\Oxid\Model\Transaction;
 
@@ -68,21 +68,10 @@ class RatepayInvoicePaymentMethodTest extends OxidEsales\TestingLibrary\UnitTest
         $this->assertFalse($this->_oPaymentMethod->isMerchantOnly());
     }
 
-    /**
-     * @dataProvider getNameProvider
-     */
-    public function testGetName($bForOxid, $sExpected)
+    public function testGetName()
     {
-        $sName = RatepayInvoicePaymentMethod::getName($bForOxid);
-        $this->assertEquals($sExpected, $sName);
-    }
-
-    public function getNameProvider()
-    {
-        return [
-            'for oxid' => [true, 'wdratepay-invoice'],
-            'not for oxid' => [false, 'ratepay-invoice'],
-        ];
+        $sName = RatepayInvoicePaymentMethod::getName();
+        $this->assertEquals('wdratepay-invoice', $sName);
     }
 
     /**
